Initial Commit - CAN Bus Sniffer
This is a CAN bus sniffer I created for observing the CAN-Bus traffic with a minimal amount of effort.
The device is implemented using a Cypress PSoC 5LP prototyping kit (CY8CKIT-059 PSoC® 5LP) is used.
This may be obtained for $10 plus shipping from http://www.cypress.com/documentation/development-kitsboards/cy8ckit-059-psoc-5lp-prototyping-kit-onboard-programmer-and
A (cheap) OBD2 reader (Does not matter if it is USB, Serial or Bluetooth) is used as a ready-made implementation of the CAN Transceiver and connector.
I would recommend keeping this component as low cost as possible, since most of its functionality will be bypassed.
For my initial unit, the CAN RX line from the CAN Transceiver of the cheap OBD2 reader is connected to P3.0.
(I bought a http://www.banggood.com/Mini-Bluetooth-OBDII-OBD2-Car-Auto-Diagnotic-Scanner-Tool-Black-p-1060352.html which contains an ON Semiconductor NCV7342 with an ELM327 clone.)
The ELM327 clone on this particular unit did not implement CAN sniffing.
The NCV7342 (Marked NV7342) has its receive line on pin 4.
I would not recommend connecting anything to the transmit line.
A ground return is provided with a via on the edge of the board.

Note: All commments provided above and in subsequent commits are advice only, and come with no warranty pursuant to the terms of the Apache license in License.txt.
Liability is expressly disclaimed for any damage to any property, including but not limited to: Computers, Development Kits, OBD-2 devices, transports (Ground, Air, Rail, Water, Space, or otherwise), or damages caused by said property to other property and people.